% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_model_sum_stat.R
\name{calc_model_sum_stat}
\alias{calc_model_sum_stat}
\title{Function to Calculate Summary Measures from Multiple Runs of the Sutherland et al.'s Model}
\usage{
calc_model_sum_stat(x, ci_level = 0.95, ci_space = "-",
  round_level = 0)
}
\arguments{
\item{x}{A list of multiple simulation runs of a singl table as produced by \code{\link{AssessBCGPolicyChange}}.}

\item{ci_level}{Numeric, specifies the confidence level to use. Defaults to 0.95.}

\item{ci_space}{Character, string to use to seperate confidence intervals, defaults to "-"}

\item{round_level}{Integer, the number of digits to which to round the summarised results. Defaults to 0}
}
\value{
A list of tables for a set of summary measures. These are; mean with CI, median, 2.5% quantile,
97.5% quantile, 25%  quantile, 75% quantile, mean, lower CI, and upper CI.
}
\description{
Function to Calculate Summary Measures from Multiple Runs of the Sutherland et al.'s Model
}
\examples{

## Run multiple model simulations
model_sims <- simulate_sutherland_model(cores = 1)

## Summarise the Total effects of ending the scheme table
calc_model_sum_stat(model_sims[["Table 5 - Total Effects of ending the schools BCG scheme"]])
}
