% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_to_data.R
\name{fit_model_to_data}
\alias{fit_model_to_data}
\title{Fit Percentage of First Year Cases}
\usage{
fit_model_to_data(model = sutherland_model,
  results = sutherland_results, ...)
}
\arguments{
\item{model}{Model function to fit, by default uses \code{\link[AssessBCGPolicyChange]{sutherland_model}}.}

\item{results}{Orginal results, by default uses \code{\link[AssessBCGPolicyChange]{sutherland_results}}}

\item{...}{Pass additional to \code{model}. See \code{\link[AssessBCGPolicyChange]{sutherland_model}} for details if
using this as the \code{model} function.}
}
\value{
The best fitting value for the percentage of cases in the first year using the sum of least squares.
}
\description{
Fit the percentage of cases in the first year using the estimation model and
original results of sutherland et al. using sum of least squares.
}
\examples{

fit_model_to_data(model = sutherland_model, results = sutherland_results)

}
