% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_annual_change_scenarios.R
\name{gen_annual_change_scenarios}
\alias{gen_annual_change_scenarios}
\title{Function to Generate Annual Change Scenarios Based on Sutherland Assumptions}
\usage{
gen_annual_change_scenarios(TB_decrease = c(0.09, 0.05, 0.01),
  cohorts_to_sim = 11, samples = 10)
}
\arguments{
\item{TB_decrease}{A numeric vector of percentage decreases in TB notifications rates (formated as 0.09 etc.).}

\item{cohorts_to_sim}{The number of cohorts to simulate. Defaults to 11}

\item{samples}{Numeric, the number of samples to generate for each scenario. Defaults to 10}
}
\value{
A dataframe with each row representing a sample, with annual decrease stored as a matrix
}
\description{
Function to Generate Annual Change Scenarios Based on Sutherland Assumptions
}
\examples{

gen_annual_change_scenarios()
}
\seealso{
TB_decrease_as_matrix
}
