% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_scenario_results.R
\name{gen_scenario_results}
\alias{gen_scenario_results}
\title{A Function to Generate Scenario Results from the Sutherland Model}
\usage{
gen_scenario_results(scenarios, parameter_samples, cores = 1, ...)
}
\arguments{
\item{scenarios}{A dataframe of scenarios, see the examples for a structural example.}

\item{parameter_samples}{A dataframe of parameter samples, see the examples for an example.}

\item{cores}{The number of cores to use, defaults to one}

\item{...}{Additional arguments passed to \code{sutherland_model}.}
}
\value{
A dataframe combining the scenarios, parameter samples, and model results
}
\description{
A Function to Generate Scenario Results from the Sutherland Model
}
\examples{

scenarios <- gen_annual_change_scenarios(TB_decrease = c(0.09, 0.05, 0.01),
cohorts_to_sim = 11, samples = 10)

parameter_samples <- sample_model_parameters(incidence_rates = sutherland_incidence_rates,
population_data = sutherland_data, samples = 10)

gen_scenario_results(scenarios, parameter_samples, cores = 1)
}
