% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_rates_uk_e.R
\docType{data}
\name{inc_rates_uk_e}
\alias{inc_rates_uk_e}
\title{Incidence Rates per 100,000 in the UK born in England}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 6 columns.}
\usage{
inc_rates_uk_e
}
\description{
A dataset containing the estimate incidence rates and associated variables for the UK born in England. Estimated using the Labour Force Survey and the Enhanced Tubercuolosis
Surveillance system. See \href{https://www.samabbott.co.uk/tbinenglanddataclean}{here}  for details.
}
\keyword{datasets}
