% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nots_ew.R
\docType{data}
\name{nots_ew}
\alias{nots_ew}
\title{Tuberculosis Notifications in England and Wales 1913-2016}
\format{A dataframe with 104 rows and 2 columns
\describe{
\item{year}{The calendar year}
\item{notifications}{The number of notified cases in any given year.}
}}
\usage{
nots_ew
}
\description{
A dataset containing the number of notifications of Tuberculosis, in any given year, in England
and Wales. Originally published by Public Health England, see \href{https://www.samabbott.co.uk/tbinenglanddataclean/reference/tb_not_ew.html}{here} for details.
}
\keyword{datasets}
