% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_annual_decrease_nots.R
\name{sample_annual_decrease_nots}
\alias{sample_annual_decrease_nots}
\title{Samples the Percentage Annual Decrease Based On Notifications}
\usage{
sample_annual_decrease_nots(notifications = NULL,
  TB_cohorts = c("15-19 years", "20-24 years", "25-29 years"),
  samples = 100, max_year = NULL, years_to_avg = 3)
}
\arguments{
\item{notifications}{A dataframe of notifications, it must contain a year variable, and the second variable
must be the number of notifications. If not specified defaults to \code{\link[AssessBCGPolicyChange]{nots_ew}}.}

\item{TB_cohorts}{A character vector indicating the names of the cohorts modelled. Defaults to 15-19 years, 20-24 years, and 25-29 years.}

\item{samples}{Numeric, the number of samples to take. Defaults to 100.}

\item{max_year}{Numeric, set as greater than the maximum year in the dataframe if interested in projection.
Defaults to \code{NULL}. If specified, then an average of the percentage decrease for a set number
of years is taken. This is specified with \code{years_to_avg}.}

\item{years_to_avg}{Numeric, the number of years over which to average the percentage decrease
when estimating the ongoing percentage decrease.}
}
\value{
A nested tibble, containing the sample number, and both a nested dataframe and nested matrix of
annual percentage changes.
}
\description{
This function samples the percentage annual decrease, using TB notifications.
The percentage annual change is first estimated using \code{prop.test}, along with the 95% confidence
intervals. Assuming that the percentage annual decrease is normally distributed multiple samples are then
taken. This is returned as a tibble, containing the nested dataframe and a nested matrix ready to be
used by \code{\link[AssessBCGPolicyChange]{sutherland_model}}.
}
\examples{

sample_annual_decrease_nots()
}
\seealso{
sample_per_decrease
}
