% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_inc_rate.R
\name{sample_inc_rate}
\alias{sample_inc_rate}
\title{Function to Sample from the Underlying Distribution of an Incidence Rate}
\usage{
sample_inc_rate(inc_rate, pop, inc_per = 1e+05, samples = 100)
}
\arguments{
\item{inc_rate}{Incidence rates formated as \code{\link[AssessBCGPolicyChange]{sutherland_incidence_rates}}.}

\item{pop}{Population estimates for each cohort as a vector.}

\item{inc_per}{The scaling factor for the incidence rate, defaults to 100,000.}

\item{samples}{Numeric, the number of samples to take. Defaults to 100.}
}
\value{
Incidence rate samples as a list, formated as a 3 by 3 matrix.
}
\description{
The number of cases are back calculated using the incidence rate and
the size of the population. Samples are then taken of notifications, assuming a poisson
distribution. Incidence rates are then calculated based on these samples. Multiple samples
of a matrix input may be taken with the output formated as a list.
}
\examples{

sample_inc_rate(Unvaccinated, sutherland_data$BCG.Unvaccinated[1:3], samples = 5)

}
