% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_incidence_rates.R
\name{sample_incidence_rates}
\alias{sample_incidence_rates}
\title{Function to sample Sutherland Incidence rates to Include Uncertainty}
\usage{
sample_incidence_rates(incidence_rates, population_data, samples = 100)
}
\arguments{
\item{incidence_rates}{Incidence rates as formated by \code{\link[AssessBCGPolicyChange]{sutherland_incidence_rates}}}

\item{population_data}{Population data as formated by  \code{\link[AssessBCGPolicyChange]{sutherland_data}}}

\item{samples}{Numeric, the number of samples to take. Defaults to 100.}
}
\value{
A list of lists. Each list contains matrices of sampled incidence rates for unvaccinated, vaccinated, tuberculin postive
and inelgible populations.
}
\description{
Function to sample Sutherland Incidence rates to Include Uncertainty
}
\examples{

sample_incidence_rates(sutherland_incidence_rates, sutherland_data, samples = 5)

}
