% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_model_parameters.R
\name{sample_model_parameters}
\alias{sample_model_parameters}
\title{Function to Sample Model Parameters}
\usage{
sample_model_parameters(incidence_rates = sutherland_incidence_rates,
  population_data = sutherland_data, samples = 100,
  as_sutherland = FALSE)
}
\arguments{
\item{incidence_rates}{Incidence rates as formated by \code{\link[AssessBCGPolicyChange]{sutherland_incidence_rates}}}

\item{population_data}{Population data as formated by  \code{\link[AssessBCGPolicyChange]{sutherland_data}}}

\item{samples}{Numeric, the number of samples to run. Defaults to 100}

\item{as_sutherland}{Logical, defaults to \code{FALSE}. If \code{TRUE} then the
original estimates used by sutherland et al. are returned. Note these are point estimates.}
}
\value{
A tibble of parameter samples
}
\description{
The serial interval is estimated as 1.44 (95% CI 1.29 to 1.63) years based on
\href{https://doi.org/10.1093/ije/dyr058}{this} study. Samples were taken by assuming a normal
distribution with a standard deviation estimated from the parameters confidence intervals.
The above study includes data on the percentage of cases in each year since infection in the
form of a figure. This appears to roughly match the original model estimate. As no other estimate
is available the point estimate from the original model is used.
}
\examples{

sample_model_parameters()

}
\seealso{
sample_incidence_rates
}
