% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_per_decrease.R
\name{sample_per_decrease}
\alias{sample_per_decrease}
\title{A Function to Sample the Annual Percentage Decrease with Uncertainty}
\usage{
sample_per_decrease(df, max_year = NULL, min_year = NULL,
  samples = 100, years_to_avg = 3)
}
\arguments{
\item{df}{A dataframe containing a year variable and with the second variable being the
one to sample from.}

\item{max_year}{Numeric, set as greater than the maximum year in the dataframe if interested in projection.
Defaults to \code{NULL}. If specified, then an average of the percentage decrease for a set number
of years is taken. This is specified with \code{years_to_avg}.}

\item{min_year}{Numeric, set as smaller than the minimum year in the dataframe if interested in backwards projection.
Defaults to \code{NULL}. If specified, then an average of the percentage decrease for a set number
of years is taken. This is specified with \code{years_to_avg}.}

\item{samples}{Numeric, the number of samples to take.}

\item{years_to_avg}{Numeric, the number of years over which to average the percentage decrease
when estimating the ongoing percentage decrease.}
}
\value{
A dataframe containing annual percentage decrease samples.
}
\description{
A Function to Sample the Annual Percentage Decrease with Uncertainty
}
\examples{

 df <- data.frame(year = c(2000:2010), measure = seq(10, 110, 10))

sample_per_decrease(df, samples = 2)

}
