% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sutherland_model.R
\name{simulate_sutherland_model}
\alias{simulate_sutherland_model}
\title{Function to run multiple simulations of the Sutherland et al. Estimation Model}
\usage{

  simulate_sutherland_model(list_annual_TB_decrease = list(TB_decrease_as_matrix(sutherland_TB_decrease),
  TB_decrease_as_matrix(sutherland_TB_decrease)),
  list_incidence_rates = list(sutherland_incidence_rates,
  sutherland_incidence_rates), list_percentage_year_one = rep(0.764, 2),
  list_sym_lag = list(2, 2), cores = 1, ...)
}
\arguments{
\item{list_annual_TB_decrease}{A list of annual TB decreases with each element as specified for
the \code{Annual.TB.Decrease.Yearly} arguement of \code{\link[AssessBCGPolicyChange]{sutherland_model}}.}

\item{list_incidence_rates}{A list of incidence rates with each element as specified for the \code{incidence_rates}
arguement of \code{\link[AssessBCGPolicyChange]{sutherland_model}}.}

\item{list_percentage_year_one}{A list of percentage secondary notifications in the first year
with each element as specified for the \code{Percentage.Year.One} arguement of
\code{\link[AssessBCGPolicyChange]{sutherland_model}}.}

\item{cores}{The number of cores to use, defaults to one}

\item{...}{Additional parameters to pass to \code{\link[AssessBCGPolicyChange]{sutherland_model}}.}
}
\value{
A list of simulations for each of the outputs of \code{\link[AssessBCGPolicyChange]{sutherland_model}}.
}
\description{
Function to run multiple simulations of the Sutherland et al. Estimation Model
}
\examples{

simulate_sutherland_model(cores = 1)

}
