% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sutherland_gen_time.R
\docType{data}
\name{sutherland_gen_time}
\alias{sutherland_gen_time}
\title{Assumed Generation time From Sutherland et al.}
\format{A dataframe with 1 variable and 1 row
\describe{
\item{sutherland_gen_time}{The assumed generation time}
}}
\usage{
sutherland_gen_time
}
\description{
A dataset containing the assumed generation time from Sutherland et al.'s original paper. See
\code{sutherland_gen_time.R} (see \code{data-raw}) and the vignette discussing recreating Sutherland et al.'s model.
}
\keyword{datasets}
