% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sutherland_model.R
\name{sutherland_model}
\alias{sutherland_model}
\title{Sutherland et al. Model to Estimate Impact of Ending the BCG School Scheme}
\usage{
sutherland_model(Data = sutherland_data,
  incidence_rates = sutherland_incidence_rates, Rates.Per = 1e+05,
  Cohort.Length = 5, Data.start = 1969,
  Annual.TB.Decrease.Yearly = TB_decrease_as_matrix(sutherland_TB_decrease),
  Percentage.Year.One = NULL, trans_params = sutherland_trans_params(),
  Sym.Lag = 2, update_chains = FALSE)
}
\arguments{
\item{Data}{Data used by Sutherland et al. analysis. See \code{\link[AssessBCGPolicyChange]{sutherland_data}}.}

\item{incidence_rates}{A list of Incidence rates (per 100,000) for the BCG vaccinated and unvaccinated population.
See \code{\link[AssessBCGPolicyChange]{sutherland_incidence_rates}} for details.}

\item{Rates.Per}{The number of people by which to produce rates, defaults to 100,000 as in Sutherland et al.}

\item{Cohort.Length}{Numeric, the lengh of time spent in each cohort, defaults to 5 years as in Sutherland et al.}

\item{Data.start}{Numeric, the year the first data point originates, defaults to 1969 as in Sutherland et al.}

\item{Annual.TB.Decrease.Yearly}{Matrix of average yearly decreases in TB with each column representing an age group (15-19, 20-24, 25-29).
Row names must be the year, starting from \code{Data.start}. Defaults to using  \code{\link[AssessBCGPolicyChange]{sutherland_TB_decrease}}.}

\item{Percentage.Year.One}{Numeric, the percentage of a generation that occur in the first year.
Fitted via, \code{\link[AssessBCGPolicyChange]{fit_model_to_data}}.}

\item{trans_params}{A funciton for estimating the transmission parameters (expected total secondary notifications,
the size of the first generation, and the average interval between a primary case and all secondary cases). Defailts to
\code{\link[AssessBCGPolicyChange]{sutherland_trans_params}}.}

\item{Sym.Lag}{Numeric, the generation time between infection and symtoms. Must be smaller or equal to the \code{Cohort.Length}.
Defaults to \code{\link[AssessBCGPolicyChange]{sutherland_gen_time}}.}

\item{update_chains}{Logical, defaults to \code{FALSE}. Should the transmission chain model be updated or used as found in Sutherland et al.}
}
\value{
A list of tables reproducing the results presented in Sutherland et al. The final table estimates the total number of additional
cases from ending the shcheme. In order to provide a complete estimate each 5 year estimate has been multiplied by the cohort length. In addition the
primary additional impacts and total secondary notifications arising from each year are also output based on paper revisions.
}
\description{
Sutherland et al. Model to Estimate Impact of Ending the BCG School Scheme
}
\examples{

sutherland_model(Percentage.Year.One = 0.764)

}
