% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sutherland_results.R
\docType{data}
\name{sutherland_results}
\alias{sutherland_results}
\title{Results for Total notifications From Sutherland et al.}
\format{A matrix with 3 rows and 6 columns. The column names indicate the estimated number of notifications in a given year,
and the rows indicate which year the schools scheme was ended.}
\usage{
sutherland_results
}
\description{
A dataset containing the estimated number of notifications for Sutherland et al.'s original paper. See
\code{sutherland_results.R} (see \code{data-raw}) and the vignette discussing recreating Sutherland et al.'s model.
}
\keyword{datasets}
