% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sutherland_TB_decrease.R
\docType{data}
\name{sutherland_TB_decrease}
\alias{sutherland_TB_decrease}
\title{Sutherland Assumption of Annual TB decrease}
\format{A vector with 3 elements. Each element represents an age group (15-19 years 20-24 years 25-29 years) and contains the projected
decrease in TB for that age group.}
\usage{
sutherland_TB_decrease
}
\description{
A dataset containing the estimated annual decrease in TB notifications from Sutherland et al.'s original paper. See
\code{sutherland_TB_decrease.R} (see \code{data-raw}) and the vignette discussing recreating Sutherland et al.'s model.
}
\keyword{datasets}
