% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sutherland_trans_params.R
\name{sutherland_trans_params}
\alias{sutherland_trans_params}
\title{Estimate Transmission chain parameters using Sutherland et al.s Approach}
\usage{

  sutherland_trans_params(Annual.TB.Decrease.Yearly = TB_decrease_as_matrix(sutherland_TB_decrease),
  Sym.Lag = sutherland_gen_time, Sutherland = TRUE, Repro.No = NULL,
  Cohort.Length = 5, update_chains = FALSE)
}
\arguments{
\item{Annual.TB.Decrease.Yearly}{Matrix of average yearly decreases in TB with each column representing an age group (15-19, 20-24, 25-29).
Row names must be the year, starting from \code{Data.start}. Defaults to using  \code{\link[AssessBCGPolicyChange]{sutherland_TB_decrease}}.}

\item{Sym.Lag}{Numeric, the generation time between infection and symtoms. Must be smaller or equal to the \code{Cohort.Length}.
Defaults to \code{\link[AssessBCGPolicyChange]{sutherland_gen_time}}.}

\item{Sutherland}{Logical, defaults to \code{TRUE}. Whether or not to use Sutherland et al.'s
assumption for estimating the basic reproduction number.}

\item{Repro.No}{If \code{Sutherland} is \code{FALSE} sets the basic reproduction number used to estimate
the transmission chain parameters.}

\item{Cohort.Length}{Numeric, the lengh of time spent in each cohort, defaults to 5 years as in Sutherland et al.}

\item{update_chains}{Logical, defaults to \code{FALSE}. Should the transmission chain model be updated or used as found in Sutherland et al.}
}
\value{
A dlist containing the expected total secondary notifications (Expected.Total.Sec.Note),
the size of the first generation (Size.First.Gen) and the average interval to all secondary cases (Avg.Int.All.Sec).
}
\description{
Estimate Transmission chain parameters using Sutherland et al.s Approach
}
\examples{

# Orginal model
sutherland_trans_params()

# Updated model
sutherland_trans_params(update_chains = TRUE)
}
