% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TB_decrease_as_matrix.R
\name{TB_decrease_as_matrix}
\alias{TB_decrease_as_matrix}
\title{Annual Decreae in TB from Single Estimate for Each Group}
\usage{
TB_decrease_as_matrix(TB_decrease, TB_cohorts = c("15-19 years",
  "20-24 years", "25-29 years"), data_start = 1969,
  projected_cohort = 9, cohort_length = 5)
}
\arguments{
\item{TB_decrease}{A named vector of percentage decreases (i.e 1\%) in TB incidence. Names should refer to age group.}

\item{TB_cohorts}{A character vector indicating the names of the cohorts modelled. Defaults to 15-19 years, 20-24 years, and 25-29 years.}

\item{data_start}{Numeric, the year the first data point originates, defaults to 1969 as in Sutherland et al.}

\item{projected_cohort}{Numeric, the number of cohort for which to simulate, defaults to 5. Combines with \code{cohort_length}
to give the length of time to simulate.}

\item{cohort_length}{Numeric, the lengh of time spent in each cohort, defaults to 5 years as in Sutherland et al.}
}
\value{
A matrix of annual TB incidence decreases as a percentage (0.01) for a given number of years and age groups
}
\description{
From a vetor of assumed annual decreases in TB, this function constructs a
matrix with yearly estimates of TB decrease
}
\examples{

TB_decrease <- c(10, 8, 11)


TB_decrease_as_matrix(TB_decrease)

}
